[#34] Move all logic, actions and outlets out of the app delegate
